Separate config mod for account controller
